refactor: journey card
